+++ Biz/Board/BoardWriteBiz.cs
@@ -11,7 +11,7 @@ namespace BoardRenual.Biz.Board
 {
     public class BoardWriteBiz
     {
-        public int InsertBoard(BoardWriteRequestModel boardWriteModel)
+        public int WriteBoard(BoardWriteRequestModel boardWriteModel)
         {
             BoardModel  boardModel = new BoardModel();
             BoardRepository boardRepository = new BoardRepository();
@@ -22,30 +22,6 @@ namespace BoardRenual.Biz.Board
             Connection connection = new Connection();
             // 객체를 넘기기
             return(boardRepository.WriteBoard(boardModel, connection));
-        }
-        public List<BoardModel> GetBoardList()
-        {
-            BoardRepository boardRepository = new BoardRepository();
-            Connection connection = new Connection();
-            return (boardRepository.GetBoardList(connection));
-        }
-        public BoardModel GetBoardDetail(int No)
-        {
-            BoardRepository boardRepository = new BoardRepository();
-            Connection connection = new Connection();
-            return (boardRepository.GetBoardDetail(connection, No));
-        }
-        public BoardModel GetBoardEmail(int No)
-        {
-            BoardRepository boardRepository = new BoardRepository();
-            Connection connection = new Connection();
-            return (boardRepository.GetBoardEmail(connection, No));
-        }
-        public void DeleteBoard(int No)
-        {
-            BoardRepository boardRepository = new BoardRepository();
-            Connection connection = new Connection();
-            boardRepository.DeleteBoard(connection, No);
         }
     }
 }
